Add camera zoom and sprinting, add timer class

+++ FusionLab/src/SDLHandler.cpp
@@ -12,7 +12,7 @@ bool compareSprites(const Sprite& sprite1, const Sprite& sprite2)
 	return uint16_t(sprite1.sprite) > uint16_t(sprite2.sprite);
 }
 
-SDLHandler::SDLHandler(const char* title, const Vector2& res, const uint8_t& cameraZoom) :
+SDLHandler::SDLHandler(const char* title, const Vector2& res, const float& cameraZoom) :
 	window(nullptr),
 	renderer(nullptr),
 	eventQueue(nullptr),
